test: SolutionCommentServiceTest 의 일부 테스트 SolutionCommentRepositoryTest로 이관
