+++ backend/src/test/java/develup/application/solution/comment/SolutionCommentServiceTest.java
@@ -5,8 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 import develup.api.exception.DevelupException;
 import develup.domain.member.Member;
 import develup.domain.member.MemberRepository;
@@ -78,43 +76,6 @@ class SolutionCommentServiceTest extends IntegrationTestSupport {
         assertThatThrownBy(() -> solutionCommentService.getComment(commentId))
                 .isInstanceOf(DevelupException.class)
                 .hasMessage("존재하지 않는 댓글입니다.");
-    }
-
-    @Test
-    @DisplayName("특정 회원이 작성한 댓글 목록을 조회한다.")
-    void getMyComments() {
-        Solution solution = createSolution();
-        Member member = createMember();
-        List<SolutionComment> solutionComments = new ArrayList<>();
-        for (int i = 0; i < 4; i++) {
-            SolutionComment solutionComment = createSolutionComment(solution, member);
-            solutionComments.add(solutionComment);
-        }
-        createSolutionComment();
-
-        List<MySolutionCommentResponse> myComments = solutionCommentService.getMyComments(member.getId());
-
-        assertThat(myComments)
-                .hasSize(solutionComments.size());
-    }
-
-    @Test
-    @DisplayName("특정 회원이 작성한 댓글 목록을 조회할 때 삭제된 댓글은 제외한다.")
-    void getMyCommentsWithDelete() {
-        Solution solution = createSolution();
-        Member member = createMember();
-        List<SolutionComment> solutionComments = new ArrayList<>();
-        for (int i = 0; i < 4; i++) {
-            SolutionComment solutionComment = createSolutionComment(solution, member);
-            solutionComments.add(solutionComment);
-        }
-        createDeletedSolutionComment(solution, member);
-        createSolutionComment();
-
-        List<MySolutionCommentResponse> myComments = solutionCommentService.getMyComments(member.getId());
-
-        assertThat(myComments)
-                .hasSize(solutionComments.size());
     }
 
     @Test
@@ -201,33 +162,6 @@ class SolutionCommentServiceTest extends IntegrationTestSupport {
         SolutionComment solutionComment = SolutionCommentTestData.defaultSolutionComment()
                 .withSolution(solution)
                 .withMember(solution.getMember())
-                .build();
-        solutionCommentRepository.save(solutionComment);
-
-        return solutionComment;
-    }
-
-    private Member createMember() {
-        Member member = MemberTestData.defaultMember().build();
-
-        return memberRepository.save(member);
-    }
-
-    private SolutionComment createSolutionComment(Solution solution, Member member) {
-        SolutionComment solutionComment = SolutionCommentTestData.defaultSolutionComment()
-                .withSolution(solution)
-                .withMember(member)
-                .build();
-        solutionCommentRepository.save(solutionComment);
-
-        return solutionComment;
-    }
-
-    private SolutionComment createDeletedSolutionComment(Solution solution, Member member) {
-        SolutionComment solutionComment = SolutionCommentTestData.defaultSolutionComment()
-                .withSolution(solution)
-                .withMember(member)
-                .withDeletedAt(LocalDateTime.now())
                 .build();
         solutionCommentRepository.save(solutionComment);
